split AccountsTable out from App
